pyproject.toml
[tool.poetry]
name = "context-python"
version = "0.3.0"
description = "Python SDK for the Context API"
authors = ["Alex Gamble <alex@woolly.ai>"]
readme = "README.md"
license = "LICENSE.md"
keywords = ["context", "api", "sdk"]
packages = [{include = "getcontext"}]

[tool.poetry.dependencies]
python = "^3.10"
azure-core = "^1.26.4"
msrest = "^0.7.1"
azure-mgmt-core = "^1.4.0"
aiohttp = "^3.8.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
